Add GitHub Action to publish `topologic_core` on PyPI
Once GitHub Release Page is published (includes changing from
draft to release), all .whl files included in the release are
collected and prepared for publishing on PyPI using GitHub Actions.

Publishing on PyPI is done with a Trusted Publisher:

Please advice at https://docs.pypi.org/trusted-publishers/creating-a-project-through-oidc/
for creating/linking PyPI projects with your GitHub repository and
this GitHub Actions Workflow (`publish.yml`).

The name of this project on PyPI is to be `topologic-core`.
Environment name is to be `release`.

Once it is done on PyPI side, you may try to publish a sample GitHub Release
(changing it from draft status to release), and expect full automatation of
PyPI publishing for the project.